card value class added, beatthedealer mostly completed
I added a CardValue class which returns the numeric value of the card.
The BeatTheDealer class has been filled out. Still need to work in the
if statements around Aces for the dealer and somehow incorporate that
for the player.

CODE DOES NOT COMPILE. Having some issues with how I declared the deck
array list in Deck.. thinking I might just have to add each card
individually because I cannot figure out how to make it take 52 at once
(keeps giving me errors for the "card","2ndcard" method.

Also, it will not accept package beatTheDealer. Apparently my directory
name is 'Beat The Dealer' and a package won't take a name in quotes..
have to change that somehow.

Took out player and dealer classes. It's too complicated to have them
separate from BeatTheDealer and I'm not able to do as much with them so
I combined the three of them.

Accept for the Ace part in BeatTheDealer, the hard parts are done! So
much debugging, though.