fixed errors, still need ace rules
code compiles! yay! for the first time!

Fixed all of the errors I was having. Just syntax errors that Aliya
knew how to fix right away so that was useful.

Turns out I deleted the entire CardValue class and replaced it with
Deck so I had two Decks and that was troublesome… almost had to rewrite
the entire class because my last commit, the first one with that class,
had it messed up.

Of course it's not running properly, though. The deck class isn't
working so I have to try to figure out what's wrong there.

Also, the code I'm seeing here that it says I'm uploading doesn't read
like what I have on my java file. But it's all highlighted so I'm not
sure if it's showing what I removed or…

Need to:
fix Deck
add ace possibilities