Finished Code, including aces, working on debugging
I added the changes that deal with aces and fixed the problem I was
having with my Deck class. I was forgetting quotes when comparing the
Strings, so it was an easy fix.

The code compiles and runs further than ever but I think I'm missing
some parenthesis because it's including more than it should on some run
throughs. There are some errors to fix in the output but they shouldn't
be too extensive.

Saving now incase I really mess it up trying to fix these output bugs.