Major Changes, added playOutPlayer()
minor bug fixes
- stuck where it asks to 'hit' or 'stint'. doesn't understand when you
input. seems like a simple fix but I haven't figured it out yet

- MAJOR
- added playOutPlayer() to beatTheDealer.BeatTheDealer. reduced that
code by 66% because I just had it typed out three times and now it's
just out once. I had to change some of the variables and move them
around but it was generally a flawless transaction! I originally tried
to make it a different class but the variables are used so intensely
between what would then have to be different classes so this was
easier. I think I might pull more code out into different methods as
well to simplify reading the code and to organize it more.  Maybe while
I'm working on that I'll figure out how to solve my little scanner
problem