Added many methods; fixed hit/stand issue
Fixed hit/stand issue by making a new instance of scanner… not sure why
it didn't work before but thanks Wyatt.

Added:
dealCards(); betting(); reset(); refillTheDeck(); cardSuit(); and
cardValue(); and noMoreGames(); to make the code easier to read and to
be able to add cards back into the deck after the game is over. makes
code more simple and collected and smaller