Minor Bug Fixes; big changes
I haven't committed in a while because I've been trying to fix the same
problems with different ways and they've hardly worked. I've made some
breakthroughs (FINALLY) though so it's worth committing. I fixed all of
the ace problems and got it through the ace loop without breaking. It
makes it through stand and hit for the most part without breaking (ace
problems fixed that)

New problems are: now it doesn't ask for another hit/stand (like have a
box... scanner not implemented right I suppose) and after the game is
won it doesn't check the booleans