Last minute touches; still need to fix ending the game
99% done! (with the original game. maybe like 85-90 with the AI)
Fixed the stand and hit issues
ended up making another method instead of manipulating one I already had
Fixed reset() so it returns all of the cards to the deck but I'm not
sure why it breaks at the end. It gets all of the cards in then out of
bounds 0,0

All I need to fix now is the ending. The game prompts the player for
whether they want to play again or not. If they say no, the game tells
them how much money they lost/won in total and ends the game. This goes
perfectly until it changes the wongame boolean to true. Then for some
reason it jumps back into one of the other methods and I cannot figure
out why.
If the player says they DO want to play again, the cards are put back
into the deck and the game restarts. But it breaks once all of the cards
are put back (while still in the replace method). Those should be really
easy fixes and once I've got them the main game will be done!

Need to clean up the code but waiting until it's done to do most of that
(a lot of extra comments and println for bug checking)