Original Game Complete (with bug checks still in)
The game is done! Except for the AI- that's still buggy and I haven't
touched it in a while. The code still has all of my bug checks in it but
I wanted to keep a copy of the finished thing with all of those checks
incase I need it.