AutoPlayer completed; W/ bug code
Fixed a couple things in BeatTheDealer including a bug that allows the
player to play again even though they have less than the needed funds.

REWROTE AutomatedPlayer and caled it AutoPlayer. There were so many
things wrong with that code I couldn't even try to fix it so I just
redid it. The automated portion of Beat The Dealer now works!

This commit has all of the bug-checking code in the AI part of the game.
Will take all of that out for the next commit.